Added a sort on the awakening materials when mapping them to the UniteMateriauxEveil
Unfortunately the order of properties is not perserved when parsing a JSON string. The method Object.getOwnPropertyNames returns the awakening material sorted by gumi_id.
As a trade-off between not doing anything about it and rewriting a JSON parser, the method convertAwakeningMaterials now sorts the array of UniteMateriauEveil it creates.
The sorting order is :
- by decreasing quantities
- by ascending gumi_ids when the quantities are equal.