Enabled comparison of all skills already present in FFCH DB.
The skill from FFCH DB was only displayed when it differed from the one in the datamining.
However the effect of the skill cannot be automatically compared, so that a manual comparison is needed.

If a skill is present in FFCH DB, both version of the skill will be displayed,
as well as the fields "puissance" and "effet".

For now, one has to open both skill panels manually.