Added the logic to post and get the enhanced status of competences
The checkbox is linked to a boolean competence.enhancedTicked.
This boolean is mapped to a string when sending a competenced to the backend, and the string is mapped to the boolean when fetching a competence.

The checkbox is displayed if the skill is not present in the DB or if the field competence.enhanced is not null.