Added a drop-down list for the Personnages in the EnhancementDisplay component
The component has now a expandable mat-card to contain the enhancement details.

There is the possibility to choose the Personnage when several of them can enhance the skill.
If there is only one Personnage, they are automatically selected and the drop-down list is disabled.

The selection is bound to the field "perso_gumi_id" in Amelioration.