Added display of competences
The EnhancementDisplay components includes a CharacterSkillsDisplay component for this task.

For enhancements of type "+1", the base skill and the new skill are diplayed.
For type "+2", the base skill, old skill and new skill are displayed.
The order is not defined, it depends on the order of the responses coming from the backend.