Added the FormuleDisplay and FormulesDisplay components
The FormuleDisplay component displays a single Formule and its status icon, given the info about its presence and correctness in FFCH DB.
The FormulesDisplay component displays both the Formule from the datamining and the Formule for the FFCH DB if necessary.
It is also responsible for fetching the images for the Formule from the datamining and for comparing the Formules.