Added the CraftableItem model and the CraftableItemsService
A CraftableItem can represent an Item, an Equipment or a Materia.
The CraftableItemCategory specify which type of craftable item is represented.

The CraftableItemService can search for CraftableItems based on their names or Gumi IDs.
It relies on the ItemsService, the EquipmentsService and the MateriasService for the actual search in the datamining.