Added the ability to fetch the ObjetCarac from the FFCH DB
The script classes.php now adds two ObjetCaracs to the Objet: One for the carac in absolute value and one for the carac in percentage.
The Objet model contains two new corresponding fields.

The ItemMapper puts placeholders for these fields if they don't exist in the received Item (in order to avoid NPE upon display).