Added Skills to Items and Competences to Objets
The Equipments and Materias have two new members: "skills" for the array of Skill Gumi IDs, and "dmSkills" for the array of Skills.
The Objet has a new field "competences" for the array of competences.
The EquipmentsService and MateriasService call the SkillsService to find the Skills in the DM.
The EquipmentMapper and MateriaMapper call the SkillMapper to create Competences out of Skills.

The ObjetDisplayComponent presents the Competences the same way as the CharacterDisplay does.

Also fixed an issue when mapping Items' effects.