Added the "stars" of an Objet
The EquipmentMapper takes the "stars" value as the "rarity" of the equipment.
The MateriaMapper takes the "stars" value as the "rarity" of the first Skill of the Materia.
The ConsumableMapper puts null for the "stars" value, because it cannot determine it.

The ObjetDisplayComponent allows changing the "stars" when the Objet is not present in the FFCH DB.
If the Objet is present in the DB, it displays the "stars" value.

The objects.php and classes.php scripts take care of storing and fetching the "stars" of the Objet.