#253: Renamed CharacterService::searchForCharacterByNameOrGumiId() into searchForCharactersByNameOrGumiId()
This method now returns an array of Characters, but it only returns the first match for now.
Modified searchForCharactersByName() in the same way.

The method to search for a character by GumiId was not changed, since the GumiId is expected to be unique.

Adapted the components using these functions and the tests.